+++ packages/elevenlabs-client/rollup.config.js
@@ -1,0 +1,6 @@
+const { libraryPreset } = require('rollup-presets');
+
+/**
+ * @type {import('rollup').RollupOptions[]}
+ */
+module.exports = libraryPreset();