+++ packages/rollup-presets/src/lib/config/create-rollup-esm-output-config.ts
@@ -1,0 +1,17 @@
+import { OutputOptions } from 'rollup';
+
+export function createRollupEsmOutputConfig(config: {
+	outputPath: string;
+	outputOptions: OutputOptions;
+}): OutputOptions {
+	const { outputOptions, outputPath } = config;
+	const { preserveModules = true } = outputOptions;
+
+	return {
+		...outputOptions,
+		[preserveModules ? 'dir' : 'file']: outputPath,
+		format: 'esm',
+		preserveModules,
+		inlineDynamicImports: !preserveModules
+	};
+}