+++ packages/rollup-presets/src/plugins/rollup-plugin-virtual-entry/index.ts
@@ -1,0 +1,21 @@
+import type { Plugin } from 'rollup';
+
+export const VIRTUAL_ENTRY_ID = 'virtual-entry';
+
+export function virtualEntryPlugin(): Plugin {
+	return {
+		name: 'virtual-entry',
+		resolveId(id) {
+			if (id === VIRTUAL_ENTRY_ID) {
+				return id;
+			}
+			return null;
+		},
+		load(id) {
+			if (id === VIRTUAL_ENTRY_ID) {
+				return 'export {}';
+			}
+			return null;
+		}
+	};
+}