+++ packages/feature-fetch/CHANGELOG.md
@@ -1,5 +1,15 @@
 # feature-fetch
 
+## 0.0.39
+
+### Patch Changes
+
+- migrated to `rollup-presets` and updated deps
+
+- Updated dependencies []:
+  - @blgc/types@0.0.16
+  - @blgc/utils@0.0.39
+
 ## 0.0.38
 
 ### Patch Changes