+++ packages/feature-form/CHANGELOG.md
@@ -1,5 +1,17 @@
 # feature-form
 
+## 0.0.44
+
+### Patch Changes
+
+- migrated to `rollup-presets` and updated deps
+
+- Updated dependencies []:
+  - validation-adapter@0.0.20
+  - feature-state@0.0.48
+  - @blgc/types@0.0.16
+  - @blgc/utils@0.0.39
+
 ## 0.0.38
 
 ### Patch Changes